+++ src/id3/v2/items/language_frame.rs
@@ -10,9 +10,8 @@ use std::hash::{Hash, Hasher};
 pub struct LanguageFrame {
 	/// The encoding of the description and comment text
 	pub encoding: TextEncoding,
-	// TODO: Make this [u8; 3] in the next major release, since it isn't always a valid UTF-8 string
 	/// ISO-639-2 language code (3 bytes)
-	pub language: String,
+	pub language: [u8; 3],
 	/// Unique content description
 	pub description: String,
 	/// The actual frame content
@@ -43,14 +42,14 @@ impl LanguageFrame {
 	pub fn as_bytes(&self) -> Result<Vec<u8>> {
 		let mut bytes = vec![self.encoding as u8];
 
-		if self.language.len() != 3 || self.language.chars().any(|c| !('a'..='z').contains(&c)) {
+		if self.language.len() != 3 || self.language.iter().any(|c| !(b'a'..=b'z').contains(&c)) {
 			return Err(ID3v2Error::new(ID3v2ErrorKind::Other(
 				"Invalid frame language found (expected 3 ascii characters)",
 			))
 			.into());
 		}
 
-		bytes.extend(self.language.as_bytes().iter());
+		bytes.extend(self.language);
 		bytes.extend(encode_text(&self.description, self.encoding, true).iter());
 		bytes.extend(encode_text(&self.content, self.encoding, false));